+++ go.mod
@@ -1,10 +1,9 @@
 module github.com/blockcypher/libgrin
 
-go 1.13
+go 1.14
 
 require (
 	github.com/btcsuite/btcd v0.20.1-beta
-	github.com/davecgh/go-spew v1.1.1
 	github.com/google/uuid v1.1.1
 	github.com/gorilla/mux v1.7.3
 	github.com/mitchellh/mapstructure v1.1.2
