+++ owner/secure_owner_api.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package example
+package owner
 
 import (
 	"crypto/ecdsa"
