+++ owner/secure_owner_api_test.go
@@ -1,0 +1,63 @@
+// Copyright 2020 BlockCypher
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http//www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package owner_test
+
+import (
+	"testing"
+
+	log "github.com/sirupsen/logrus"
+)
+
+func checkError(err error) {
+	if err != nil {
+		log.Fatal(err)
+	}
+}
+
+func TestReal(t *testing.T) {
+	// commenting this since this can't be done on CI for now
+	//url := "http://127.0.0.1:3420/v3/owner"
+
+	//ownerAPI := owner.NewSecureOwnerAPI(url)
+	//err := ownerAPI.Init()
+	//assert.Nil(t, err, "init owner api err was not nil")
+
+	//err = ownerAPI.Open(nil, "")
+	//assert.Nil(t, err, "open wallet err was not nil")
+
+	//nodeHeight, err := ownerAPI.NodeHeight()
+	//assert.Nil(t, err, "node height err was not nil")
+	//assert.NotNil(t, nodeHeight)
+
+	////torConfig := libwallet.TorConfig{
+	////	UseTorListener: true,
+	////	SocksProxyAddr: "127.0.0.1:59050",
+	////	SendConfigDir:  ".",
+	////}
+	////if err := ownerAPI.SetTorConfig(torConfig); err != nil {
+	////	assert.Error(t, err)
+	////}
+
+	//fresh, summary, err := ownerAPI.RetrieveSummaryInfo(true, 10)
+	//assert.Nil(t, err, "retrieve summary err was not nil")
+
+	//assert.True(t, fresh, "refresh from node should be true")
+
+	//log.Infof("%+v\n", summary)
+
+	//if err := ownerAPI.Close(nil); err != nil {
+	//	assert.Error(t, err)
+	//}
+}
